refactor: Create editor utils file
